pyproject.toml
[project]
name = "microservice-template"
version = "0.1.0"
description = ""
authors = [
    {name = "andre",email = "andrepedro2004@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "grpcio-tools (>=1.70.0,<2.0.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "odmantic (>=1.0.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic-extra-types (>=2.10.2,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "strawberry-graphql[debug-server,opentelemetry] (>=0.262.5,<0.263.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.52b0,<0.53)",
    "opentelemetry-instrumentation-grpc (>=0.52b0,<0.53)",
    "opentelemetry-instrumentation-sqlalchemy (>=0.52b0,<0.53)",
    "opentelemetry-sdk (>=1.31.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.31.0,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
black = "^25.1.0"
isort = "^6.0.1"
sqlalchemy-stubs = "^0.4"
types-protobuf = "^5.29.1.20250208"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
